added new changes

+++ PostOffice.Service/Services/BagService.cs
@@ -1,10 +1,13 @@
+using PostOffice.Common.Requests;
+using PostOffice.Repository.Entities;
+
 namespace PostOffice.Service.Services;
 public interface IBagService
 {
     Task<IEnumerable<BagResponse>> GetAllAsync();
     Task<BagResponse> GetByIdAsync(int id);
     Task CreateAsync(BagRequest model);
-    Task UpdateAsync(BagUpdateRequest model);
+    Task UpdateAsync(BagRequest model);
     Task DeleteAsync(int id);
 }
 
@@ -14,11 +17,15 @@ public class BagService : IBagService
 
     private readonly IBagRepository _bagRepository;
 
-    public BagService(IBagRepository bagRepository,
-        IMapper mapper)
+    private readonly IParcelRepository _parcelRepository;
+
+    private readonly IShipmentRepository _shipmentRepository;
+    public BagService(IMapper mapper, IBagRepository bagRepository, IParcelRepository parcelRepository, IShipmentRepository shipmentRepository)
     {
-        _bagRepository = bagRepository;
         _mapper = mapper;
+        _bagRepository = bagRepository;
+        _parcelRepository = parcelRepository;
+        _shipmentRepository = shipmentRepository;
     }
 
     public async Task<IEnumerable<BagResponse>> GetAllAsync()
@@ -45,22 +52,87 @@ public class BagService : IBagService
 
     public async Task CreateAsync(BagRequest model)
     {
+        if (model == null) throw new ArgumentNullException("model");
+
         // map model to new Bag object
         var bag = _mapper.Map<Bag>(model);
 
         // save Bag
-        await _bagRepository.CreateAsync(bag);
+        var bagId = await _bagRepository.CreateAsync(bag);
+
+        if(bagId > 0 && model.ParcelRequests != null && model.ParcelRequests.Count > 0)
+        {
+            foreach (var parcelRequest in model.ParcelRequests)
+            {
+                var parcel = _mapper.Map<Parcel>(parcelRequest);
+
+                parcel.BagId = bagId;
+                await _parcelRepository.CreateAsync(parcel);
+            }
+        }
     }
 
-    public async Task UpdateAsync(BagUpdateRequest model)
+    public async Task UpdateAsync(BagRequest model)
     {
+        if (model == null) throw new ArgumentNullException("model");
+
+        if(!IsValid(model.ShipmentId ?? 0))
+            return;
+
         // copy model to Bag and update
         var bag = _mapper.Map<Bag>(model);
         await _bagRepository.UpdateAsync(bag);
+
+        if (model.ParcelRequests != null && model.ParcelRequests.Count > 0)
+        {
+            var parcels = await _parcelRepository.GetAllByBagIdAsync(model.BagId);
+
+            var newParcels = model.ParcelRequests.Where(q => q.ParcelId == 0).ToList();
+
+            foreach (var parcelRequest in newParcels)
+            {
+                var parcel = _mapper.Map<Parcel>(parcelRequest);
+                parcel.BagId = model.BagId;
+                await _parcelRepository.CreateAsync(parcel);
+            }
+
+            var excludeParcels = parcels.Where(q => model.ParcelRequests.Any(p => p.ParcelId > 0 && p.ParcelId != q.ParcelId)).ToList();
+
+            foreach (var parcel in excludeParcels)
+            {
+                parcel.BagId = null;
+                await _parcelRepository.UpdateAsync(parcel);
+            }
+        }
     }
 
     public async Task DeleteAsync(int id)
     {
+        var model = await GetByIdAsync(id);
+
+        if (!IsValid(model?.ShipmentId ?? 0))
+            return;
+
+        var parcels = await _parcelRepository.GetAllByBagIdAsync(id);
+
+        foreach (var parcel in parcels)
+        {
+            await _parcelRepository.DeleteAsync(parcel.ParcelId);
+        }
+
         await _bagRepository.DeleteAsync(id);
+    }
+
+    public bool IsValid(int shipmentId)
+    {
+        if(shipmentId > 0)
+        {
+            var shipment = _shipmentRepository.GetByIdAsync(shipmentId);
+
+            if (shipment != null && (Status)shipment.Status == Status.Finalized)
+                return false;
+        }
+
+        return true;
     }
 }
