+++ PostOffice.Service/Services/ShipmentService.cs
@@ -1,3 +1,6 @@
+using PostOffice.Common.Requests;
+using PostOffice.Repository.Entities;
+
 namespace PostOffice.Service.Services;
 public interface IShipmentService
 {
@@ -14,11 +17,16 @@ public class ShipmentService : IShipmentService
 
     private readonly IShipmentRepository _shipmentRepository;
 
-    public ShipmentService(IShipmentRepository shipmentRepository,
-        IMapper mapper)
+    private readonly IBagRepository _bagRepository;
+
+    private readonly IParcelRepository _parcelRepository;
+
+    public ShipmentService(IMapper mapper, IShipmentRepository shipmentRepository, IBagRepository bagRepository, IParcelRepository parcelRepository)
     {
-        _shipmentRepository = shipmentRepository;
         _mapper = mapper;
+        _shipmentRepository = shipmentRepository;
+        _bagRepository = bagRepository;
+        _parcelRepository = parcelRepository;
     }
 
     public async Task<IEnumerable<ShipmentResponse>> GetAllAsync()
@@ -45,25 +53,95 @@ public class ShipmentService : IShipmentService
 
     public async Task CreateAsync(ShipmentRequest model)
     {
+        if (model == null) throw new ArgumentNullException("model");
+
         // map model to new Shipment object
         model.FlightDate = Convert.ToDateTime(model.FlightDate);
         var shipment = _mapper.Map<Shipment>(model);
 
         // save Shipment
-        await _shipmentRepository.CreateAsync(shipment);
+        var shipmentId = await _shipmentRepository.CreateAsync(shipment);
+
+        if (shipmentId > 0 && model.Bags != null && model.Bags.Count > 0)
+        {
+            foreach (var bagRequest in model.Bags)
+            {
+                var bag = _mapper.Map<Bag>(bagRequest);
+
+                bag.ShipmentId = shipmentId;
+                var bagId = await _bagRepository.CreateAsync(bag);
+            }
+        }
     }
 
     public async Task UpdateAsync(ShipmentUpdateRequest model)
     {
+        if (model == null) throw new ArgumentNullException("model");
+
+        if (!IsValid(model.ShipmentId))
+            return;
+
         // copy model to Shipment and update
         var shipment = _mapper.Map<Shipment>(model);
         model.FlightDate = Convert.ToDateTime(model.FlightDate);
         await _shipmentRepository.UpdateAsync(shipment);
+
+        if (model.Bags != null && model.Bags.Count > 0)
+        {
+            var bags = await _bagRepository.GetAllByShipmentIdAsync(model.ShipmentId);
+
+            var newBags = model.Bags.Where(q => q.BagId == 0).ToList();
+
+            foreach (var bagRequest in newBags)
+            {
+                var bag = _mapper.Map<Bag>(bagRequest);
+                bag.ShipmentId = model.ShipmentId;
+                await _bagRepository.CreateAsync(bag);
+            }
+
+            var excludeBags = bags.Where(q => model.Bags.Any(p => p.BagId > 0 && p.BagId != q.BagId)).ToList();
+
+            foreach (var bag in excludeBags)
+            {
+                bag.ShipmentId = null;
+                await _bagRepository.UpdateAsync(bag);
+            }
+        }
     }
 
     public async Task DeleteAsync(int id)
     {
+        if (!IsValid(id))
+            return;
+
+        var bags = await _bagRepository.GetAllByShipmentIdAsync(id);
+
+        foreach (var bag in bags)
+        {
+            var parcels = await _parcelRepository.GetAllByBagIdAsync(bag.BagId);
+
+            foreach (var parcel in parcels)
+            {
+                await _parcelRepository.DeleteAsync(parcel.ParcelId);
+            }
+
+            await _bagRepository.DeleteAsync(bag.BagId);
+        }
+
         await _shipmentRepository.DeleteAsync(id);
+    }
+
+    public bool IsValid(int shipmentId)
+    {
+        if (shipmentId > 0)
+        {
+            var shipment = _shipmentRepository.GetByIdAsync(shipmentId);
+
+            if (shipment != null && (Status)shipment.Status == Status.Finalized)
+                return false;
+        }
+
+        return true;
     }
 
 }
